+++ FrontEnd/FrontEnd.csproj
@@ -23,6 +23,10 @@
     </PackageReference>
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\Domain\Domain.csproj" />
+  </ItemGroup>
+
 <!--  <ItemGroup>-->
 <!--    <Folder Include="DTO" />-->
 <!--  </ItemGroup>-->
