+++ FrontEnd/Pages/FleetPages/FleetView.razor.cs
@@ -1,6 +1,7 @@
 ﻿using Microsoft.AspNetCore.Components;
 using System;
 using System.Threading.Tasks;
+using Domain.Entities;
 using FlyingPizza.Domain.Entities;
 
 
@@ -8,14 +9,14 @@ namespace FlyingPizza.Pages.FleetPages
 {
     partial class FleetView : ComponentBase
     {
-        public DroneModel[] Fleet = null;
+        public DroneFields[] Fleet = null;
         public int size;
         public Boolean connection = false;
         protected override async Task OnInitializedAsync()
         {
             try
             {
-                Fleet = await restPoint.Get<DroneModel[]>("http://localhost:8080/Fleet/?sort={badgeNumber:1}");
+                Fleet = await restPoint.Get<DroneFields[]>("http://localhost:8080/Fleet/?sort={badgeNumber:1}");
                 size = Fleet.Length;
                 connection = true;
             }
@@ -23,7 +24,7 @@ namespace FlyingPizza.Pages.FleetPages
             }         
         }
 
-        public async Task GoToDrone(DroneModel drone)
+        public async Task GoToDrone(DroneFields drone)
         {
             globalData.currDrone = drone;
             await dialogService.OpenAsync<DetailedDrone>("View Drone");
