+++ SimulatedDrone/Drone.cs
@@ -8,44 +8,8 @@ using static System.Decimal;
 
 namespace DroneSimulator
 {
-    public enum DroneState
+    public class Drone : DroneFields
     {
-        Ready,
-        Delivering,
-        Returning,
-        Dead,
-        Charging
-    }
-
-    public class Drone
-    {
-        // Radius of the Earth used in calculating distance 
-        private const int EarthRadius = 6371;
-
-        // 20 MPH as meters per second
-        private const double DroneSpeed = 0.0089408;
-        
-        // Number of milliseconds to wait before updating Drone status
-        private const int DroneUpdateInterval = 2000;
-
-        // I don't think this makes sense but it's working...
-        private const decimal StepSize = DroneUpdateInterval / 1000.0m * (decimal)DroneSpeed;
-        
-        // The unique ID of this drone which is stored in the database
-        private string Id { get; }
-
-        // The point representing the pizza restaurant
-        public GeoLocation Home { get; set; }
-
-        // The current position of the drone
-        public GeoLocation Location { get; set; }
-
-        // The desired position of the drone
-        public GeoLocation Destination { get; set; }
-
-        // Current state of the drone
-        private DroneState State { get; set; }
-
         // Gateway for communication with the dispatcher
         private readonly IDispatcherGateway _dispatcher;
 
@@ -53,33 +17,33 @@ namespace DroneSimulator
         public Drone(string id, GeoLocation home, IDispatcherGateway dispatcher)
         {
             Id = id;
-            Location = home;
+            CurrentLocation = home;
             Destination = home;
             State = DroneState.Ready;
             _dispatcher = dispatcher;
-            Home = home;
+            HomeLocation = home;
         }
         
         // Return an array of Geolocations representing a drone's delivery route
         public GeoLocation[] GetRoute()
         {
-            if (Home.Equals(Destination))
+            if (HomeLocation.Equals(Destination))
             {
                 throw new ArgumentException(
                     "Destination cannot be the same as the Delivery station!");
             }
 
-            var distance = Haversine(ToDouble(Home.Latitude), ToDouble(Home.Longitude),
+            var distance = Haversine(ToDouble(HomeLocation.Latitude), ToDouble(HomeLocation.Longitude),
                 ToDouble(Destination.Latitude), ToDouble(Destination.Longitude));
 
             var numberOfLocations = (int)Math.Floor((decimal)distance / StepSize);
 
             Console.WriteLine($"need to travel {distance} km, step_size={StepSize}, num locations={numberOfLocations}");
             // Latitude distance to get to destination
-            var xStep = (Destination.Longitude - Home.Longitude) / numberOfLocations;
+            var xStep = (Destination.Longitude - HomeLocation.Longitude) / numberOfLocations;
             
             // Longitude distance to get to destination
-            var yStep = (Destination.Latitude - Home.Latitude) / numberOfLocations;
+            var yStep = (Destination.Latitude - HomeLocation.Latitude) / numberOfLocations;
             
             // LINQ yields all the points (except possibly the last one) along the route, one unit apart.
             var route = Enumerable.Range(0, numberOfLocations - 1)
@@ -141,18 +105,18 @@ namespace DroneSimulator
         // Send an Location update to DispatcherGateway
         private void UpdateLocation(GeoLocation location)
         {
-            Location = location;
+            CurrentLocation = location;
             _dispatcher.PutDroneState(
                 new PutStatusDto
                 {
                     BadgeNumber = Id,
-                    Location = Location
+                    Location = CurrentLocation
                 });
         }
 
         public override string ToString()
         {
-            return $"Drone:{{Id:{Id},Location:{Location},Destination:{Destination},State:{State}}}";
+            return $"Drone:{{Id:{Id},Location:{CurrentLocation},Destination:{Destination},State:{State}}}";
         }
         
         // Helper function for Haversine formula readability
@@ -167,6 +131,16 @@ namespace DroneSimulator
             double lat2, double lon2) =>
         EarthRadius * 2 * Math.Asin(
             Math.Sqrt(SinSquared(lat2 - lat1)) + SinSquared((lon2 - lon1) * Math.Cos(ToRadians(lat1)) * Math.Cos(ToRadians(lat2))));
+        
+        public override bool Equals(object o)
+        {
+            if (o == null || o.GetType() != GetType()) return false;
+            Drone oo = (Drone) o;
+            return oo.BadgeNumber == BadgeNumber &&
+                   oo.CurrentLocation.Equals(CurrentLocation) &&
+                   oo.Destination.Equals(Destination) &&
+                   oo.State.Equals(State);
+        }
     }
 }
